requirements.txt
# Core PDF processing
PyMuPDF==1.23.19
pdfplumber==0.10.4

# OCR capabilities
pytesseract==0.3.10
opencv-python==4.8.1.78

# For PaddleOCR (alternative OCR engine)
paddlepaddle==2.4.2
paddleocr==2.7.3

# For document generation
python-docx==1.1.0

# Data processing
pandas==2.1.4
numpy==1.24.3

# Utilities
pillow==10.0.1